discord/requirements.txt
nextcord==2.6.0
yt-dlp==2023.11.16
python-dotenv==1.0.0
pynacl==1.5.0
